requirements.txt
flask==2.3.3
flask-cors==4.0.0
speechrecognition==3.10.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydub==0.25.1
faster-whisper==1.1.1
gunicorn==21.2.0 